Break apart monolithic profile sheets
