+++ xcode/Subconscious/Shared/Components/Common/Profile/Sheets/FollowNewUserFormSheet.swift
@@ -1,14 +1,14 @@
 //
-//  FollowUserFormView.swift
-//  Subconscious (iOS)
+//  FollowNewUserFormSheet.swift
+//  Subconscious
 //
-//  Created by Ben Follington on 2/27/23.
+//  Created by Ben Follington on 3/11/2023.
 //
 
-import os
 import SwiftUI
 import ObservableStore
 import CodeScanner
+import os
 
 struct FollowNewUserFormSheetView: View {
     var store: ViewStore<FollowNewUserFormSheetModel>
@@ -34,7 +34,7 @@ struct FollowNewUserFormSheetView: View {
         store.send(.attemptFollow)
     }
     
-    func onCancel() { 
+    func onCancel() {
         store.send(.dismissSheet)
     }
     
@@ -45,7 +45,7 @@ struct FollowNewUserFormSheetView: View {
                     FollowUserFormView(
                         store: store.viewStore(
                             get: \.form,
-                            tag: FollowUserFormCursor.tag
+                            tag: FollowNewUserFormCursor.tag
                         )
                     )
                     
@@ -136,6 +136,53 @@ enum FollowNewUserFormSheetAction {
 
 typealias FollowNewUserFormSheetEnvironment = AppEnvironment
 
+// MARK: Cursors
+
+struct FollowNewUserFormCursor: CursorProtocol {
+    typealias Model = FollowNewUserFormSheetModel
+    typealias ViewModel = FollowUserFormModel
+
+    static func get(state: Model) -> ViewModel {
+        state.form
+    }
+    
+    static func set(state: Model, inner: ViewModel) -> Model {
+        var model = state
+        model.form = inner
+        return model
+    }
+    
+    static func tag(_ action: ViewModel.Action) -> Model.Action {
+        .form(action)
+    }
+}
+
+struct FollowNewUserFormSheetCursor: CursorProtocol {
+    typealias Model = UserProfileDetailModel
+    typealias ViewModel = FollowNewUserFormSheetModel
+
+    static func get(state: Model) -> ViewModel {
+        state.followNewUserFormSheet
+    }
+    
+    static func set(state: Model, inner: ViewModel) -> Model {
+        var model = state
+        model.followNewUserFormSheet = inner
+        return model
+    }
+    
+    static func tag(_ action: ViewModel.Action) -> Model.Action {
+        switch action {
+        case .attemptFollow:
+            return UserProfileDetailAction.attemptFollowNewUser
+        case .dismissSheet:
+            return UserProfileDetailAction.presentFollowNewUserFormSheet(false)
+        default:
+            return UserProfileDetailAction.followNewUserFormSheet(action)
+        }
+    }
+}
+
 // MARK: Model
 struct FollowNewUserFormSheetModel: ModelProtocol {
     typealias Action = FollowNewUserFormSheetAction
@@ -161,7 +208,7 @@ struct FollowNewUserFormSheetModel: ModelProtocol {
     ) -> Update<Self> {
         switch (action) {
         case .form(let action):
-            return FollowUserFormCursor.update(
+            return FollowNewUserFormCursor.update(
                 state: state,
                 action: action,
                 environment: ()
@@ -227,210 +274,5 @@ struct FollowNewUserFormSheetModel: ModelProtocol {
             // Handled by FollowNewUserFormSheetCursor.tag
             return Update(state: state)
         }
-    }
-}
-
-// MARK: Cursors
-
-struct FollowUserFormCursor: CursorProtocol {
-    typealias Model = FollowNewUserFormSheetModel
-    typealias ViewModel = FollowUserFormModel
-
-    static func get(state: Model) -> ViewModel {
-        state.form
-    }
-    
-    static func set(state: Model, inner: ViewModel) -> Model {
-        var model = state
-        model.form = inner
-        return model
-    }
-    
-    static func tag(_ action: ViewModel.Action) -> Model.Action {
-        .form(action)
-    }
-}
-
-struct FollowNewUserFormSheetCursor: CursorProtocol {
-    typealias Model = UserProfileDetailModel
-    typealias ViewModel = FollowNewUserFormSheetModel
-
-    static func get(state: Model) -> ViewModel {
-        state.followNewUserFormSheet
-    }
-    
-    static func set(state: Model, inner: ViewModel) -> Model {
-        var model = state
-        model.followNewUserFormSheet = inner
-        return model
-    }
-    
-    static func tag(_ action: ViewModel.Action) -> Model.Action {
-        switch action {
-        case .attemptFollow:
-            return UserProfileDetailAction.attemptFollowNewUser
-        case .dismissSheet:
-            return UserProfileDetailAction.presentFollowNewUserFormSheet(false)
-        default:
-            return UserProfileDetailAction.followNewUserFormSheet(action)
-        }
-    }
-}
-
-
-// MARK: Inner FollowUserForm
-struct FollowUserFormView: View {
-    var store: ViewStore<FollowUserFormModel>
-    
-    var petnameCaption: String {
-        store.state.failFollowMessage ?? "Lowercase letters, numbers and dashes only."
-    }
-    
-    var body: some View {
-        Section(header: Text("User To Follow")) {
-            HStack(alignment: .top) {
-                Image(systemName: "key")
-                    .foregroundColor(.accentColor)
-                
-                ValidatedFormField(
-                    placeholder: "DID",
-                    field: store.viewStore(
-                        get: \.did,
-                        tag: FollowUserFormAction.didField
-                    ),
-                    caption: String(
-                        localized: "e.g. did:key:z6MkmCJAZansQ3p1Qwx6wrF4c64yt2rcM8wMrH5Rh7DGb2K7"
-                    ),
-                    axis: .vertical
-                )
-                .lineLimit(12)
-                .textInputAutocapitalization(.never)
-                .disableAutocorrection(true)
-            }
-            
-            HStack(alignment: .top) {
-                Image(systemName: "at")
-                    .foregroundColor(.accentColor)
-                
-                ValidatedFormField(
-                    placeholder: "petname",
-                    field: store.viewStore(
-                        get: \.petname,
-                        tag: FollowUserFormAction.petnameField
-                    ),
-                    caption: petnameCaption
-                )
-                .lineLimit(1)
-                .textInputAutocapitalization(.never)
-                .disableAutocorrection(true)
-            }
-        }
-    }
-}
-
-// MARK: Actions
-
-enum FollowUserFormAction: Equatable {
-    case didField(FormFieldAction<String>)
-    case petnameField(FormFieldAction<String>)
-    case reset
-}
-
-// MARK: Model
-
-struct FollowUserFormModel: ModelProtocol {
-    typealias Action = FollowUserFormAction
-    typealias Environment = ()
-    
-    var did: FormField<String, Did> = FormField(
-        value: "",
-        validate: Self.validateDid
-    )
-    
-    var petname: FormField<String, Petname.Name> = FormField(
-        value: "",
-        validate: Self.validatePetname
-    )
-    
-    var failFollowMessage: String? = nil
-    
-    static func validateDid(key: String) -> Did? {
-        Did(key)
-    }
-
-    static func validatePetname(petname: String) -> Petname.Name? {
-        Petname.Name(petname)
-    }
-    
-    static func update(
-        state: Self,
-        action: Action,
-        environment: Environment
-    ) -> Update<Self> {
-        switch action {
-        case .didField(let action):
-            return DidFieldCursor.update(
-                state: state,
-                action: action,
-                environment: FormFieldEnvironment()
-            )
-        case .petnameField(let action):
-            return PetnameFieldCursor.update(
-                state: state,
-                action: action,
-                environment: FormFieldEnvironment()
-            )
-        case .reset:
-            var model = state
-            model.failFollowMessage = nil
-            return update(
-                state: model,
-                actions: [
-                    .didField(.reset),
-                    .petnameField(.reset)
-                ],
-                environment: environment
-            )
-        }
-    }
-}
-
-// MARK: Cursors
-
-private struct PetnameFieldCursor: CursorProtocol {
-    typealias Model = FollowUserFormModel
-    typealias ViewModel = FormField<String, Petname.Name>
-
-    static func get(state: Model) -> ViewModel {
-        state.petname
-    }
-    
-    static func set(state: Model, inner: ViewModel) -> Model {
-        var model = state
-        model.petname = inner
-        return model
-    }
-    
-    static func tag(_ action: ViewModel.Action) -> Model.Action {
-        FollowUserFormAction.petnameField(action)
-    }
-}
-
-private struct DidFieldCursor: CursorProtocol {
-    typealias Model = FollowUserFormModel
-    typealias ViewModel = FormField<String, Did>
-
-    static func get(state: Model) -> ViewModel {
-        state.did
-    }
-    
-    static func set(state: Model, inner: ViewModel) -> Model {
-        var model = state
-        model.did = inner
-        return model
-    }
-    
-    static func tag(_ action: ViewModel.Action) -> Model.Action {
-        FollowUserFormAction.didField(action)
     }
 }
