+++ xcode/Subconscious/Shared/Components/Common/Profile/Sheets/FollowUserFormView.swift
@@ -1,0 +1,167 @@
+//
+//  FollowUserFormView.swift
+//  Subconscious (iOS)
+//
+//  Created by Ben Follington on 2/27/23.
+//
+
+import os
+import SwiftUI
+import ObservableStore
+
+// MARK: Inner FollowUserForm
+struct FollowUserFormView: View {
+    var store: ViewStore<FollowUserFormModel>
+    
+    var petnameCaption: String {
+        store.state.failFollowMessage ?? "Lowercase letters, numbers and dashes only."
+    }
+    
+    var body: some View {
+        Section(header: Text("User To Follow")) {
+            HStack(alignment: .top) {
+                Image(systemName: "key")
+                    .foregroundColor(.accentColor)
+                
+                ValidatedFormField(
+                    placeholder: "DID",
+                    field: store.viewStore(
+                        get: \.did,
+                        tag: FollowUserFormAction.didField
+                    ),
+                    caption: String(
+                        localized: "e.g. did:key:z6MkmCJAZansQ3p1Qwx6wrF4c64yt2rcM8wMrH5Rh7DGb2K7"
+                    ),
+                    axis: .vertical
+                )
+                .lineLimit(12)
+                .textInputAutocapitalization(.never)
+                .disableAutocorrection(true)
+            }
+            
+            HStack(alignment: .top) {
+                Image(systemName: "at")
+                    .foregroundColor(.accentColor)
+                
+                ValidatedFormField(
+                    placeholder: "petname",
+                    field: store.viewStore(
+                        get: \.petname,
+                        tag: FollowUserFormAction.petnameField
+                    ),
+                    caption: petnameCaption
+                )
+                .lineLimit(1)
+                .textInputAutocapitalization(.never)
+                .disableAutocorrection(true)
+            }
+        }
+    }
+}
+
+// MARK: Actions
+
+enum FollowUserFormAction: Equatable {
+    case didField(FormFieldAction<String>)
+    case petnameField(FormFieldAction<String>)
+    case reset
+}
+
+// MARK: Model
+
+struct FollowUserFormModel: ModelProtocol {
+    typealias Action = FollowUserFormAction
+    typealias Environment = ()
+    
+    var did: FormField<String, Did> = FormField(
+        value: "",
+        validate: Self.validateDid
+    )
+    
+    var petname: FormField<String, Petname.Name> = FormField(
+        value: "",
+        validate: Self.validatePetname
+    )
+    
+    var failFollowMessage: String? = nil
+    
+    static func validateDid(key: String) -> Did? {
+        Did(key)
+    }
+
+    static func validatePetname(petname: String) -> Petname.Name? {
+        Petname.Name(petname)
+    }
+    
+    static func update(
+        state: Self,
+        action: Action,
+        environment: Environment
+    ) -> Update<Self> {
+        switch action {
+        case .didField(let action):
+            return DidFieldCursor.update(
+                state: state,
+                action: action,
+                environment: FormFieldEnvironment()
+            )
+        case .petnameField(let action):
+            return PetnameFieldCursor.update(
+                state: state,
+                action: action,
+                environment: FormFieldEnvironment()
+            )
+        case .reset:
+            var model = state
+            model.failFollowMessage = nil
+            return update(
+                state: model,
+                actions: [
+                    .didField(.reset),
+                    .petnameField(.reset)
+                ],
+                environment: environment
+            )
+        }
+    }
+}
+
+// MARK: Cursors
+
+private struct PetnameFieldCursor: CursorProtocol {
+    typealias Model = FollowUserFormModel
+    typealias ViewModel = FormField<String, Petname.Name>
+
+    static func get(state: Model) -> ViewModel {
+        state.petname
+    }
+    
+    static func set(state: Model, inner: ViewModel) -> Model {
+        var model = state
+        model.petname = inner
+        return model
+    }
+    
+    static func tag(_ action: ViewModel.Action) -> Model.Action {
+        FollowUserFormAction.petnameField(action)
+    }
+}
+
+private struct DidFieldCursor: CursorProtocol {
+    typealias Model = FollowUserFormModel
+    typealias ViewModel = FormField<String, Did>
+
+    static func get(state: Model) -> ViewModel {
+        state.did
+    }
+    
+    static func set(state: Model, inner: ViewModel) -> Model {
+        var model = state
+        model.did = inner
+        return model
+    }
+    
+    static func tag(_ action: ViewModel.Action) -> Model.Action {
+        FollowUserFormAction.didField(action)
+    }
+}
